proofchain01/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Image Processing & Hashing
Pillow==10.2.0
ImageHash==4.3.1

# Blockchain - Soroban/Stellar
stellar-sdk==9.6.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1